Simple creep that shoot back and kill you really fast implemented.

+++ SharedContent/EntityModels/CombatStats.cs
@@ -14,7 +14,7 @@ namespace SharedContent
         public int Resistance;
 
         public float Damage;
-        public float DamageType;
+        public int DamageType;
 
         public CombatStats clone()
         {
